Postgres: Ignore 'internal' namespaces during Load
This change fixed a number of odd bugs that can occur when using
testfixtures with `timescaledb` - a Postgres extension for time-series
data. Timescale partitions tables automatically to enable
high-throughput and query optimization. Timescale tables are placed into
a few namespaces which all start with an underscore (_), which are
considered private, off-limits by the extension. This commit causes
testfixtures to ignore tables in namespaces that start with an
underscore.

This change may break other applications that generate tables in
similarly named namespaces.